fix: TreesInvariants is divided into several parts: AVLTreeInvariants, BSTreeInvariants, RBTreeInvariants.
