+++ lib/src/test/kotlin/binarySearchTree/BSTreeTest.kt
@@ -3,7 +3,6 @@ package binarySearchTree
 import org.junit.jupiter.api.*
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import treeInvariants.TreesInvariants
 import bstrees.model.trees.binarySearch.BSNode
 import bstrees.model.trees.binarySearch.BSTree
 import kotlin.random.Random
@@ -15,7 +14,7 @@ class BSTreeTest {
     private lateinit var keyValue: List<Pair<Int, Int>>
     private lateinit var bigKeyValue: List<Pair<Int, Int>>
     private val tree = BSTree<Int, Int>()
-    private val treeChecker = TreesInvariants<Int, Int, BSNode<Int, Int>>()
+    private val treeChecker = BSTreeInvariants<Int, Int, BSNode<Int, Int>>()
 
     @BeforeAll
     fun prepareNodes() {
